// apps/gq/src/app/shared/services/rest/f-pricing/f-pricing.paths.enum.ts
export enum FPricingPaths {
  PATH_QUOTATION_DETAILS = 'quotation-details',
  PATH_F_PRICING = 'f-pricing',
  PATH_F_PRICING_CALCULATIONS = 'final-price-calculation',
  PATH_F_PRICING_FEATURE_COMPARISON = 'feature-comparison',
  PATH_COMPARABLE_K_NUMBER_TRANSACTIONS = 'comparable-k-number-transactions',
}